asdf

+++ torch_baselines/common/losses.py
@@ -37,7 +37,7 @@ class CategorialDistributionLoss(_Loss):
         self.batch_size = batch_size
         offset = torch.linspace(0, (self.batch_size - 1) * categorial_bar_n, self.batch_size)
         offset = offset.unsqueeze(dim=1) 
-        self.offset = offset.expand(self.batch_size, categorial_bar_n).int() # I believe this is to(device)
+        self.offset = offset.expand(self.batch_size, categorial_bar_n) # I believe this is to(device)
 
 
     def forward(self, input_distribution: Tensor, next_distribution: Tensor, next_categorial_bar: Tensor) -> Tensor:
@@ -49,7 +49,7 @@ class CategorialDistributionLoss(_Loss):
             C51_U = C51_b.ceil().int()
             C51_L[ (C51_U > 0) * (C51_L == C51_U)] -= 1
             C51_U[ (C51_L < (self.categorial_bar_n - 1)) * (C51_L == C51_U)] += 1
-            self.offset = self.offset.to(next_distribution)
+            self.offset = self.offset.to(next_distribution).int()
             target_distribution = input_distribution.new_zeros(self.batch_size, self.categorial_bar_n) # Returns a Tensor of size size filled with 0. same dtype
             target_distribution.view(-1).index_add_(0, (C51_L + self.offset).view(-1), (next_distribution * (C51_U.float() - C51_b)).view(-1))
             target_distribution.view(-1).index_add_(0, (C51_U + self.offset).view(-1), (next_distribution * (C51_b - C51_L.float())).view(-1))
